[PRN-298321: Updated POST endpoint for Exporter and Reprocessor + wrote more unit tests]

+++ EPRN.Portal/Areas/Exporter/Controllers/PRNSController.cs
@@ -243,6 +243,9 @@ namespace EPRN.Portal.Areas.Exporter.Controllers
         [ActionName(Routes.Areas.Actions.PRNS.DeleteDraft)]
         public async Task<IActionResult> DeleteDraftPrn(DeleteDraftPrnViewModel viewModel)
         {
+            if (viewModel == null)
+                return BadRequest();
+
             await _prnService.DeleteDraftPrn(viewModel);
 
             return RedirectToAction("ViewDraftPRNS", new { viewModel.Id }); //TODO: This needs to go to the View Draft PRNs page when it's developed
